refactor: split compressOrDecompress to two functions
